Fix spec, output needs to be a map
